update link "Community Network Manual by APC" to include what APC stands for
"by APC" is an abbreviation for "Association for Progressive Communications". Adding this information helps avoid confusion. APC is a widely known (to computer and network folks) company that provides (among other things) battery backup and rackmount cabinets for datacenters. As this list will be of interest to network engineers (even aspiring hobby ones), at least in the US, APC will be misunderstood.